qt-cpp:natvis-test: extract debug/build helpers and document NatVis snapshot pipeline
