+++ qt-cpp/test/helper.mts
@@ -10,6 +10,14 @@ import * as os from 'os';
 import { isDeepStrictEqual } from 'util';
 import { delay } from 'qt-lib';
 
+// ---------------------------------------------------------------------------
+// Shared debug logging + sandbox lifecycle for this test file
+// ---------------------------------------------------------------------------
+export const DEBUG = process.env.QT_TEST_DEBUG === '1';
+export const dlog = (...args: unknown[]) => {
+  if (DEBUG) console.log(...args);
+};
+
 /**
  * Mocha lifecycle wiring for a shared Sinon sandbox.
  * - creates a sandbox once
